Get rid of extension manager impl
